Avance finalizacion ingreso de productos

+++ database/migrations/2025_04_14_143810_create_producto_sucursals_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger("sucursal_id");
             $table->unsignedBigInteger("producto_id");
-            $table->unsignedBigInteger("stock_actual");
+            $table->double("stock_actual");
             $table->timestamps();
         });
     }
